# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "EpisodicMemoryListResponse": {
        "description": "Paginated episode list \u2014 newest first within the (since, until) window.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EpisodicMemoryResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "EpisodicMemoryListResponse",
        "type": "object"
      },
      "EpisodicMemoryResponse": {
        "description": "One (s, a, r, s') episode.",
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "consumed_by_sleep_run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consumed By Sleep Run Id"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "meta_data": {
            "additionalProperties": true,
            "title": "Meta Data",
            "type": "object"
          },
          "next_state_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next State Key"
          },
          "order_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Id"
          },
          "pair_id": {
            "format": "uuid",
            "title": "Pair Id",
            "type": "string"
          },
          "reward": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Reward",
            "type": "string"
          },
          "state_key": {
            "title": "State Key",
            "type": "string"
          }
        },
        "required": [
          "id",
          "pair_id",
          "state_key",
          "action",
          "reward"
        ],
        "title": "EpisodicMemoryResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PairListResponse": {
        "description": "Paginated list payload \u2014 mirror of the X-Total-Count header in the body.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PairSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "PairListResponse",
        "type": "object"
      },
...
      "PairSummary": {
        "description": "Slim representation for list endpoints.",
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "timeframe": {
            "title": "Timeframe",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "id",
          "name",
          "symbol",
          "timeframe",
          "status"
        ],
        "title": "PairSummary",
        "type": "object"
      },
...
      "QTableListResponse": {
        "description": "Paginated Q-Table list \u2014 newest version first.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/QTableListItem"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "QTableListResponse",
        "type": "object"
      },
      "QTableResponse": {
        "description": "Full Q-Table version, including the ``table_data`` JSONB.",
        "properties": {
          "alpha_normal": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Alpha Normal",
            "type": "string"
          },
          "alpha_special": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Alpha Special",
            "type": "string"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "created_by_sleep_run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By Sleep Run Id"
          },
          "episode_count": {
            "title": "Episode Count",
            "type": "integer"
          },
          "gamma": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Gamma",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "pair_id": {
            "format": "uuid",
            "title": "Pair Id",
            "type": "string"
          },
          "table_data": {
            "additionalProperties": true,
            "title": "Table Data",
            "type": "object"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "pair_id",
          "version",
          "alpha_normal",
          "alpha_special",
          "gamma",
          "episode_count"
        ],
        "title": "QTableResponse",
        "type": "object"
      },
...
      "SemanticMemoryListResponse": {
        "description": "Active rule list \u2014 ordered newest-first.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SemanticMemoryResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "SemanticMemoryListResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Aether API",
    "version": "0.0.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/pairs": {
      "get": {
        "description": "List the caller's pairs. Ignores any client-supplied user_id.\n\nPagination is offset/limit. See ``pair_repository`` docstring for\nthe planned cursor migration path.",
        "operationId": "list_pairs_api_pairs_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 20,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 10000,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Pairs",
        "tags": [
          "pairs"
        ]
      },
      "post": {
        "description": "Create a pair owned by the current user.\n\nStatus is hard-coded to ``inactive`` (the DDL default). The caller\nmust drive it through the lifecycle endpoints to reach any other\nstate \u2014 this prevents accidentally publishing an active pair on\na typo and matches the state machine's \"new rows enter via inactive\nedge\" invariant.",
        "operationId": "create_pair_api_pairs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PairCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Pair",
        "tags": [
          "pairs"
        ]
      }
    },
...
    "/api/pairs/{pair_id}/episodic-memory": {
      "get": {
        "description": "Return episodes for ``pair_id`` filtered by time window + state_key.\n\nDefaults:\n  * ``since`` = ``now - 7 days``\n  * ``until`` = ``now``\n\n404 on cross-tenant. ``total`` reflects the page length (we don't\nissue a second COUNT query \u2014 pagination is offset/limit anyway).",
        "operationId": "list_episodic_memory_api_pairs__pair_id__episodic_memory_get",
        "parameters": [
          {
            "in": "path",
            "name": "pair_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Pair Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          },
          {
            "in": "query",
            "name": "until",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Until"
            }
          },
          {
            "in": "query",
            "name": "state_key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 120,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State Key"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 100000,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EpisodicMemoryListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Episodic Memory",
        "tags": [
          "learning"
        ]
      }
    },
...
    "/api/pairs/{pair_id}/q-tables": {
      "get": {
        "description": "List Q-Table versions for ``pair_id`` (newest version first).\n\n404 if the pair does not belong to the caller (existence\nnon-disclosure). Cross-tenant calls never see a 403.\n\n``total`` reflects ALL versions for the pair (not the page size).",
        "operationId": "list_q_tables_api_pairs__pair_id__q_tables_get",
        "parameters": [
          {
            "in": "path",
            "name": "pair_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Pair Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 10000,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QTableListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Q Tables",
        "tags": [
          "learning"
        ]
      }
    },
    "/api/pairs/{pair_id}/q-tables/{version}": {
      "get": {
        "description": "Return a single Q-Table version (full ``table_data`` JSONB included).\n\n404 if the version does not exist OR the pair is cross-tenant \u2014\nwe don't differentiate, by design.",
        "operationId": "get_q_table_version_api_pairs__pair_id__q_tables__version__get",
        "parameters": [
          {
            "in": "path",
            "name": "pair_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Pair Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "title": "Version",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QTableResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Q Table Version",
        "tags": [
          "learning"
        ]
      }
    },
    "/api/pairs/{pair_id}/semantic-memory": {
      "get": {
        "description": "Return semantic rules for ``pair_id``.\n\n``active`` defaults to True. The repository surface (``list_active``)\nonly exposes active rows \u2014 passing ``active=false`` returns an empty\nlist rather than the supersession history (the long-running history\nsurface is intentionally out of scope for Phase 9).\n\n404 on cross-tenant.",
        "operationId": "list_semantic_memory_api_pairs__pair_id__semantic_memory_get",
        "parameters": [
          {
            "in": "path",
            "name": "pair_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Pair Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "rule_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 40,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rule Type"
            }
          },
          {
            "in": "query",
            "name": "active",
            "required": false,
            "schema": {
              "default": true,
              "title": "Active",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SemanticMemoryListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Semantic Memory",
        "tags": [
          "learning"
        ]
      }
    },
...
  }
}